+++ DevOps.Primitives.CSharp/TypeDeclaration.cs
@@ -61,7 +61,7 @@ namespace DevOps.Primitives.CSharp
         [ProtoMember(18)]
         public ModifierList ModifierList { get; set; }
         [ProtoMember(19)]
-        public int? ModifierListId { get; set; }
+        public byte? ModifierListId { get; set; }
 
         [ProtoMember(20)]
         public Namespace Namespace { get; set; }
